Split ResponsesServer into focused modules
Move the persist-before-terminal contract to terminal.ts, bounded
request-body reading to body.ts, the foreground collect/stream execution
modes to foreground.ts, the detached background run driver to
background-driver.ts, and the id-addressed routes (GET, DELETE, /cancel,
/input_items) to resource-routes.ts. server.ts keeps routing, response
creation and the id-claim lifecycle; the moved code is unchanged apart
from receiving its dependencies as explicit parameters. No public API
changes.
